third_party/WebKit/LayoutTests/typedcssom/inlinestyle/inlineStylePropertyMap_getProperties.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

test(function() {
  testElement.style = '';
  assert_array_equals(testElement.styleMap.getProperties(), []);
}, "getProperties returns an empty list when no properties have been set");

test(function() {
  testElement.style = '';
  testElement.styleMap.set('width', new CSSSimpleLength(10, 'px'));
  assert_array_equals(testElement.styleMap.getProperties(), ['width']);
}, "getProperties returns the name of a property if it is set");

test(function() {
  testElement.styleMap.set('width', new CSSSimpleLength(10, 'px'));
  assert_array_equals(testElement.styleMap.getProperties(), ['width']);

  testElement.styleMap.get('height');
  assert_array_equals(testElement.styleMap.getProperties(), ['width']);
}, "Accessing another property doesn't add a spurious result");

test(function() {
  testElement.styleMap.set('width', new CSSSimpleLength(10, 'px'));
  assert_array_equals(testElement.styleMap.getProperties(), ['width']);

  testElement.styleMap.delete('width');
  assert_array_equals(testElement.styleMap.getProperties(), []);
}, "property name does not appear in result after deletion");

test(function() {
  testElement.styleMap.set('width', new CSSSimpleLength(10, 'px'));
  assert_array_equals(testElement.styleMap.getProperties(), ['width']);

  testElement.styleMap.set('border-top-width', new CSSSimpleLength(10, 'px'));
  var result = testElement.styleMap.getProperties();
  // TODO(meade): The spec should describe an order for this.
  assert_equals(result.length, 2);
  assert_true(result.indexOf('width') >= 0);
  assert_true(result.indexOf('border-top-width') >= 0);
}, "getProperties returns multiple properties if they are set.");

test(function() {
  testElement.style = '';
  testElement.style.setProperty('--my-custom-property', '5px');

  assert_array_equals(testElement.styleMap.getProperties(), ['--my-custom-property']);
}, "getProperties returns expected values for custom properties");

test(function() {
    testElement.style.cssText = "@apply --foo";
    assert_array_equals(testElement.styleMap.getProperties(), ['@apply']);
}, "getProperties returns expected values when @apply is used");

test(function() {
    testElement.style.cssText = "@apply --foo; @apply --bar;";
    assert_array_equals(testElement.styleMap.getProperties(), ['@apply']);
}, "getProperties returns only one @apply when multiple things are applied");

</script>
